labels: Move CIDR related code into pkg/labels/cidr
[ upstream commit 08935a65faf0fd47ad454efbebafc1e7a82d77ee ]
